+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 ### Features:
  - Github Actions
  - Github issue templates
- - Basic API for cards, decks, tricks and games
+ - Basic API for cards, decks, tricks and deals
  - Rotterdam-style trump suit mechanics
+ - Amsterdam-style trump suit mechanics